+++ frontend/src/components/Tooltip/Tooltip.styles.tsx
@@ -1,6 +1,6 @@
 import { CSSProperties } from 'react';
 import styled from '@emotion/styled';
-import { Z_INDEX } from '@/constants/elementOrder';
+import { Z_INDEX } from '@/constants/style';
 
 const StyledContainer = styled.div`
   position: relative;
